requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
Django==3.1.6
django-cors-headers==3.7.0
django-extra-fields==3.0.2
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
greenlet==1.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
pathlib-mate==1.0.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
SQLAlchemy==1.4.22
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
uszipcode==0.2.6
